Rename to rspec-rails-swagger
This is really an extension of rspec-rails so make that clear in the name.

Fixes #5
