+++ lib/rspec/rails/swagger/version.rb
@@ -1,0 +1,10 @@
+module RSpec
+  module Rails
+    # Version information for RSpec Swagger.
+    module Swagger
+      module Version
+        STRING = '0.1.0'
+      end
+    end
+  end
+end
