--- lib/rspec/swagger/version.rb
@@ -1,8 +1,0 @@
-module RSpec
-  # Version information for RSpec Swagger.
-  module Swagger
-    module Version
-      STRING = '0.1.0'
-    end
-  end
-end
