+++ spec/rspec/rails/swagger/version_spec.rb
@@ -1,0 +1,7 @@
+require 'swagger_helper'
+
+RSpec.describe RSpec::Rails::Swagger do
+  it "loads" do
+    expect(RSpec::Rails::Swagger::Version::STRING).to eq '0.1.0'
+  end
+end
