--- spec/rspec/swagger/version_spec.rb
@@ -1,7 +1,0 @@
-require 'swagger_helper'
-
-RSpec.describe RSpec::Swagger do
-  it "loads" do
-    expect(RSpec::Swagger::Version::STRING).to eq '0.1.0'
-  end
-end
