+++ spec/spec_helper.rb
@@ -1,4 +1,4 @@
-require 'rspec/swagger'
+require 'rspec/rails/swagger'
 
 # This file was generated by the `rspec --init` command. Conventionally, all
 # specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
